feat(profile): extract Mission Control domain
