+++ extensions/mission-control/mission-control-actions/domain/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""Mission Control Suite-owned domain implementation."""
+
+from . import mission_bundle, mission_control
+
+__all__ = ["mission_bundle", "mission_control"]
